Implementation of DBSCAN Clustering
A nSort-like microclustering based on a DBSCAN algorithm was implemented in the read_g4 function. A temporal grouping with a time scale of 10 ns is performed previous to the spatial clustering to prevent the merging of events clearly separated in time. For each cluster the number of quanta is generated with NEST whereas the interaction type is chosen individually for each cluster based on the particle types of the clustered events.